Add a `preserve_metadata` option
This is boolean flag in config, which defaults to False. If set to True,
then the vendor task will *not* remove `.dist-info` and `.egg-info`
files.

The motivation is that some libraries use `pkg_resources` or
`import_metadata` to inspect locally installed packages, e.g for plugin
loading. Previously, the metadata needed for those APIs was deleted.

This change allows a user to preserve the metadata in the vendoring dir.
They will still need to add that dir to sys.path for it to work, but at
least they have the option to do that now.